Revert "* optimizing bfast match process.  Removing some unecessary file"
This reverts commit 467fe1cebbd38c909e71706d5ffdb83861fac022.

Nils Homer: this optimization in fact makes things slower.  The changes
being reverted got rid of the N-way merge when using multiple indexes
by using the previous output as input to the next index.  In this way, the
index results would be automatically merged. Most likely explanations
include that more had to be read (bytes off disk), and/or that the merge
at each new index search was too expensive and it was better just to offload
the merge to the end.